# docs/openapi/chainmove.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChainMove API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://example.invalid"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "chainmove_session"
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-paystack-signature"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": true
            }
          }
        },
        "required": [
          "message"
        ],
        "additionalProperties": true
      },
      "MoneyMinor": {
        "type": "object",
        "required": [
          "currency",
          "amountMinor"
        ],
        "properties": {
          "currency": {
            "type": "string",
            "description": "ISO 4217 currency code."
          },
          "amountMinor": {
            "type": "integer",
            "description": "Minor units such as kobo or cents."
          }
        }
      }
    }
  },
  "paths": {
    "/api/payments/initialize": {
      "post": {
        "tags": [
          "payments"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "amountNgn": {
                    "type": "number"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "authorization_url": {
                          "type": "string"
                        },
                        "access_code": {
                          "type": "string"
                        },
                        "reference": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": true
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pools/{poolId}/invest": {
      "post": {
        "tags": [
          "investments"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "poolId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amountNgn": {
                    "type": "number"
                  },
                  "txRef": {
                    "type": "string"
                  }
                },
                "required": [
                  "amountNgn"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        true
                      ]
                    },
                    "investment": {
                      "type": "object",
                      "properties": {
                        "poolId": {
                          "type": "string"
                        },
                        "userId": {
                          "type": "string"
                        },
                        "amountNgn": {
                          "type": "number"
                        },
                        "ownershipUnits": {
                          "type": "number"
                        },
                        "ownershipBps": {
                          "type": "number"
                        },
                        "txRef": {
                          "type": "string"
                        },
                        "poolStatus": {
                          "type": "string",
                          "enum": [
                            "OPEN",
                            "FUNDED",
                            "CLOSED"
                          ]
                        },
                        "currentRaisedNgn": {
                          "type": "number"
                        },
                        "targetAmountNgn": {
                          "type": "number"
                        },
                        "investorCount": {
                          "type": "number"
                        },
                        "userBalanceNgn": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "poolId",
                        "userId",
                        "amountNgn",
                        "ownershipUnits",
                        "ownershipBps",
                        "txRef",
                        "poolStatus",
                        "currentRaisedNgn",
                        "targetAmountNgn",
                        "investorCount",
                        "userBalanceNgn"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "success",
                    "investment"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/webhook": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "security": [
          {
            "webhookSignature": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event": {
                    "type": "string"
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": true
                  }
                },
                "required": [
                  "event",
                  "data"
                ],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "received": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  }
}